refactor: split responsability ChatManager/ChatContainer/ChatRoom
